Refactor -  데이터처리 컴포넌트 별도 분리하며 Inner컴포넌트 호출 거리 최소화
