--- components/categories/Categories.type.ts
@@ -1,5 +1,0 @@
-import { ContItems } from 'apis/models/Categories.type';
-
-export default interface CategoriesProps {
-  tabInfo: ContItems[];
-}
