+++ components/categories/index.styled.ts
