+++ components/categories/index.type.ts
@@ -1,0 +1,8 @@
+import { CategoriesAPIType, ContItems } from 'apis/models/Categories.type';
+
+export default interface CategoriesProps {
+  data: {
+    categoryInfo: CategoriesAPIType;
+    tabInfo: ContItems[];
+  };
+}
