+++ pages/categories/[category1Id].tsx
@@ -1,16 +1,18 @@
 import React from 'react';
-import { AppLayout, TabMenu, Category, Spinner } from 'components/common';
-import * as S from 'components/categories/Categories.styled';
-import { useRouter } from 'next/router';
-import { get } from 'apis/requestAPIs/categories';
-import useSWR from 'swr';
-import CategoriesProps from 'components/categories/Categories.type';
-import { GetStaticProps } from 'next';
+import Categories from 'components/categories';
+import { AppLayout, Spinner } from 'components/common';
 import Error from 'pages/_error';
-import { ItemThumbnail } from 'components/common';
+import { get } from 'apis/requestAPIs/categories';
+import { GetStaticProps } from 'next';
+import { useRouter } from 'next/router';
+import useSWR from 'swr';
 import { ContItems } from 'apis/models/Categories.type';
 
-const Categories = ({ tabInfo }: CategoriesProps) => {
+interface CategoriesPageProps {
+  tabInfo: ContItems[];
+}
+
+const CategoriesPage = ({ tabInfo }: CategoriesPageProps) => {
   const router = useRouter();
   const { category1Id } = router.query;
 
@@ -18,46 +20,17 @@ const Categories = ({ tabInfo }: CategoriesProps) => {
     get.categories(category1Id),
   );
 
-  const getDisCountItem = (name: string, innerItem: ContItems) => ({
-    brand: name,
-    href: `/items/${innerItem.id}`,
-    imageUrl: innerItem.imageUrl,
-    itemName: innerItem.name,
-    discountRate: innerItem.discountRate,
-    minSellingPrice: innerItem?.minSellingPrice,
-    originalPrice: innerItem.originalPrice,
-  });
-
   return (
     <AppLayout title={categoryInfo?.conCategory1.name}>
-      {categoryInfo ? (
-        <>
-          <TabMenu menuData={tabInfo} tabType="category" />
-          {categoryInfo.conCategory1.name === '땡철이' ? (
-            // categoryInfo?.conCategory1.conCategory2s 브랜드 목록
-            categoryInfo.conCategory1.conCategory2s.map(brand =>
-              brand.conItems?.map(conItem => {
-                const discountItem = getDisCountItem(brand.name, conItem);
-                return (
-                  <S.Box key={discountItem.itemName}>
-                    <ItemThumbnail {...discountItem} />
-                  </S.Box>
-                );
-              }),
-            )
-          ) : (
-            <S.List>
-              {categoryInfo.conCategory1.conCategory2s.map(item => (
-                <Category page="brands" key={item.id} item={item} />
-              ))}
-            </S.List>
-          )}
-        </>
-      ) : error ? (
-        <Error statusCode={404} />
-      ) : (
-        <Spinner />
-      )}
+      {(() => {
+        if (error) {
+          return <Error statusCode={error.response.status} />;
+        } else if (!categoryInfo) {
+          return <Spinner />;
+        } else {
+          return <Categories data={{ categoryInfo, tabInfo }} />;
+        }
+      })()}
     </AppLayout>
   );
 };
@@ -93,4 +66,4 @@ export const getStaticPaths = async () => {
   };
 };
 
-export default Categories;
+export default CategoriesPage;
